content_moderation_openenv/server/requirements.txt
openenv[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
#fastapi==0.115.0
#uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
#openai>=1.40.0
pytest==8.3.3
httpx==0.27.2

